+++ web/core/components/workspace-notifications/index.ts
@@ -1,2 +1,3 @@
 export * from "./notification-app-sidebar-option";
 export * from "./sidebar";
+export * from "./root";
