+++ web/core/components/workspace-notifications/sidebar/notification-card/index.ts
@@ -1,3 +1,2 @@
-export * from "./root";
 export * from "./item";
 export * from "./options";
